# frontend/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sagara Mission Control API
  version: 1.0.0
  description: |
    Canonical Sagara Mission Control API Specification (Contract V1).
    Frozen for future FastAPI backend implementation.
    Adheres strictly to snake_case naming, ISO 8601 UTC timestamps, opaque string IDs,
    pessimistic mutation locking, explicit idempotency keys, and UNKNOWN != ZERO semantics.
servers:
  - url: http://localhost:8000/api/v1
    description: Local Mission Control Backend
  - url: https://mission-control.sagara.local/api/v1
    description: Production Internal Gateway
paths:
...
  /profiles:
    get:
      summary: List Registered Agent Profiles
      operationId: listProfiles
      responses:
        '200':
          description: List of agent profile manifests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileDto'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /profiles/{profile_id}:
    get:
      summary: Get Profile by ID
      operationId: getProfileById
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profile definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /agents:
    get:
      summary: List Agent Projections
      operationId: listAgents
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: profile_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of agent runtime projections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentDto'
  /agents/{agent_id}:
    get:
      summary: Get Agent Projection by ID
      operationId: getAgentById
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent projection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /skills:
    get:
      summary: List Skill Projections
      operationId: listSkills
      parameters:
        - name: registration
          in: query
          required: false
          schema:
            type: string
        - name: health
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Skills list with multi-dimensional status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillDto'
  /skills/{skill_id}:
    get:
      summary: Get Skill Projection by ID
      operationId: getSkillById
      parameters:
        - name: skill_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDto'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /delegations:
    get:
      summary: List Runtime Subtask Delegations
      operationId: listDelegations
      parameters:
        - name: agent_id
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of subtask delegations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DelegationDto'
  /delegations/{delegation_id}:
    get:
      summary: Get Delegation by ID
      operationId: getDelegationById
      parameters:
        - name: delegation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delegation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelegationDto'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /approvals:
    get:
      summary: List Human Approval Gates
      operationId: listApprovals
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: risk
          in: query
          schema:
            type: string
        - name: action_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Approvals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApprovalDto'
  /approvals/{approval_id}:
    get:
      summary: Get Approval Gate by ID
      operationId: getApprovalById
      parameters:
        - name: approval_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approval gate details
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /approvals/{approval_id}/approve:
    post:
      summary: Authorize Action
      operationId: approveAction
      parameters:
        - name: approval_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionInputDto'
      responses:
        '200':
          description: Action approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDto'
        '409':
          description: Race condition (e.g. APPROVAL_ALREADY_RESOLVED, APPROVAL_EXPIRED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
  /approvals/{approval_id}/reject:
    post:
      summary: Deny Action
      operationId: rejectAction
      parameters:
        - name: approval_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionInputDto'
      responses:
        '200':
          description: Action rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDto'
        '409':
          description: Race condition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
...
  /artifacts:
    get:
      summary: List Task Output Artifacts
      operationId: listArtifacts
      parameters:
        - name: task_id
          in: query
          schema:
            type: string
        - name: session_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Artifacts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtifactDto'
  /artifacts/{artifact_id}:
    get:
      summary: Get Artifact Metadata
      operationId: getArtifactById
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artifact projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactDto'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    ApiErrorDto:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: APPROVAL_ALREADY_RESOLVED
            message:
              type: string
              example: This approval gate has already been decided.
            details:
              type: object
              nullable: true
            correlation_id:
              type: string
              nullable: true
              example: corr-20260909-001
    ProfileDto:
      type: object
      required:
        - id
        - name
        - enabled
      properties:
        id:
          type: string
          example: prof-sagara-lead
        name:
          type: string
          example: Sagara Lead Orchestrator
        role:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        enabled:
          type: boolean
        memory_namespace:
          type: string
          nullable: true
        allowed_skills:
          type: array
          items:
            type: string
          nullable: true
    AgentDto:
      type: object
      required:
        - id
        - definition
        - runtime
        - capabilities
      properties:
        id:
          type: string
          example: agent-alpha
        definition:
          type: object
          required:
            - name
            - enabled
          properties:
            name:
              type: string
            role:
              type: string
              nullable: true
            description:
              type: string
              nullable: true
            enabled:
              type: boolean
            memory_namespace:
              type: string
              nullable: true
        runtime:
          type: object
          required:
            - state
            - confidence
          properties:
            state:
              type: string
              enum:
                - ACTIVE
                - IDLE
                - RECENTLY_ACTIVE
                - AWAITING_APPROVAL
                - DEGRADED
                - ERROR
                - OFFLINE
                - UNKNOWN
                - CONFIGURATION_INCOMPLETE
            confidence:
              type: string
              enum:
                - CONFIRMED
                - INFERRED
                - UNKNOWN
                - STALE
            last_activity_at:
              type: string
              format: date-time
              nullable: true
            session_count:
              type: integer
              nullable: true
            active_delegations:
              type: integer
              nullable: true
            current_session_id:
              type: string
              nullable: true
            current_task_id:
              type: string
              nullable: true
            model:
              type: string
              nullable: true
        capabilities:
          type: object
          properties:
            total:
              type: integer
              nullable: true
            healthy:
              type: integer
              nullable: true
            degraded:
              type: integer
              nullable: true
            missing:
              type: integer
              nullable: true
        usage:
          type: object
          nullable: true
          properties:
            input_tokens:
              type: integer
              nullable: true
            output_tokens:
              type: integer
              nullable: true
            estimated_cost_usd:
              type: number
              nullable: true
    SkillDto:
      type: object
      required:
        - id
        - name
        - category
        - registration
        - installation
        - health
        - execution
      properties:
        id:
          type: string
          example: skill-code-executor
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
        registration:
          type: string
          enum: [REGISTERED, UNREGISTERED]
        installation:
          type: string
          enum: [INSTALLED, MISSING, UNKNOWN]
        health:
          type: string
          enum: [HEALTHY, DEGRADED, MISSING, UNKNOWN]
        execution:
          type: string
          enum: [NOT_OBSERVED, REQUESTED, EXECUTION_UNKNOWN, OBSERVED_ACTIVE, COMPLETED, FAILED]
...
    ApprovalDto:
      type: object
      required:
        - id
        - state
        - risk
        - action_type
        - title
        - requested_at
      properties:
        id:
          type: string
          example: appr-01
        state:
          type: string
          enum: [PENDING, APPROVED, REJECTED, EXPIRED, CANCELLED, FAILED]
        risk:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        action_type:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        reason_required:
          type: boolean
          nullable: true
        task_id:
          type: string
          nullable: true
        agent_id:
          type: string
          nullable: true
        requested_at:
          type: string
          format: date-time
        preview:
          type: object
          nullable: true
          properties:
            summary:
              type: string
              nullable: true
            sensitive_fields:
              type: array
              items:
                type: string
              nullable: true
    ApprovalDecisionInputDto:
      type: object
      properties:
        reason:
          type: string
          nullable: true
...
    DelegationDto:
      type: object
      required:
        - id
        - task_title
        - state
        - started_at
      properties:
        id:
          type: string
          example: del-01
        parent_session_id:
          type: string
          nullable: true
        target_agent_id:
          type: string
          nullable: true
        task_title:
          type: string
        state:
          type: string
          enum: [QUEUED, CLAIMED, RUNNING, COMPLETED, FAILED, CANCELLED, UNKNOWN]
        worker_pid:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
...
    ArtifactDto:
      type: object
      required:
        - id
        - name
        - created_at
      properties:
        id:
          type: string
        task_id:
          type: string
          nullable: true
        session_id:
          type: string
          nullable: true
        name:
          type: string
        media_type:
          type: string
          nullable: true
        size_bytes:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [AVAILABLE, PROCESSING, MISSING, EXPIRED, UNKNOWN]
...
  responses:
...
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'
...
    ServiceUnavailable:
      description: Upstream runtime or service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorDto'